requirements.txt
alembic==1.9.4
altair==4.2.2
analytics-python==1.4.0
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
attrs==22.2.0
Babel==2.11.0
backcall==0.2.0
backoff==1.10.0
beautifulsoup4==4.11.1
bleach==6.0.0
blinker==1.5
cachetools==5.3.0
catboost==1.1.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
cloudpickle==2.2.1
cmaes==0.9.1
colorama==0.4.6
colorlog==6.7.0
comm==0.1.2
contourpy==1.0.7
crispy-bootstrap4==2022.1
cryptography==39.0.2
cvxopt==1.3.0
cvxpy==1.3.0
cycler==0.11.0
databricks-cli==0.17.4
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
Django==4.1.7
django-allauth==0.52.0
django-appconf==1.0.5
django-bootstrap-v5==1.0.11
django-bootstrap4==22.3
django-cookie-consent==0.3.1
django-crispy-forms==2.0
django-oauth==1.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
docker==6.0.1
ecos==2.0.12
entrypoints==0.4
evidently==0.2.6
exceptiongroup==1.1.0
executing==1.2.0
fancyimpute==0.7.0
fastjsonschema==2.16.2
flake8==6.0.0
Flask==2.2.3
fonttools==4.38.0
fqdn==1.5.1
future==0.18.3
gitdb==4.0.10
GitPython==3.1.30
graphviz==0.20.1
gunicorn==20.1.0
h5py==3.8.0
hyperopt==0.2.7
idna==3.4
imbalanced-learn==0.10.1
importlib-metadata==5.2.0
iniconfig==2.0.0
ipykernel==6.20.2
ipython==8.8.0
ipython-genutils==0.2.0
ipywidgets==8.0.4
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter_client==7.4.9
jupyter_core==5.1.5
jupyter_server==2.1.0
jupyter_server_terminals==0.4.4
jupyterlab==3.5.3
jupyterlab-pygments==0.2.2
jupyterlab-spellchecker==0.7.2
jupyterlab-widgets==3.0.5
jupyterlab_server==2.19.0
kiwisolver==1.4.4
knnimpute==0.1.0
lightgbm==3.3.5
llvmlite==0.39.1
Mako==1.2.4
Markdown==3.4.1
markdown-it-py==2.1.0
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mccabe==0.7.0
mdurl==0.1.2
mistune==2.0.4
mlflow==2.1.1
monotonic==1.6
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
networkx==3.0
nltk==3.8.1
nose==1.3.7
notebook==6.5.2
notebook_shim==0.2.2
numba==0.56.4
numpy==1.23.5
oauthlib==3.2.2
optuna==3.1.0
osqp==0.6.2.post8
packaging==22.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==2.6.2
plotly==5.13.0
pluggy==1.0.0
prometheus-client==0.16.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==10.0.1
pybind11==2.10.3
pycodestyle==2.10.0
pycparser==2.21
pydeck==0.8.0
pyflakes==3.0.1
Pygments==2.14.0
PyJWT==2.6.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
pytest==7.2.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-json-logger==2.0.4
python3-openid==3.2.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
pyzmq==25.0.0
qdldl==0.1.5.post3
qtconsole==5.4.0
QtPy==2.3.0
querystring-parser==1.2.4
regex==2022.10.31
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.1
scikit-learn==1.2.1
scipy==1.10.0
scs==3.2.2
seaborn==0.12.2
semver==2.13.0
Send2Trash==1.8.0
shap==0.41.0
six==1.16.0
slicer==0.0.7
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.46
sqlparse==0.4.3
stack-data==0.6.2
statsmodels==0.13.5
streamlit==1.17.0
tabulate==0.9.0
tenacity==8.1.0
termcolor==2.2.0
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
Unidecode==1.3.6
uri-template==1.2.0
urllib3==1.26.14
validators==0.20.0
watchdog==2.3.1
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.0
Werkzeug==2.2.3
widgetsnbextension==4.0.5
xgboost==1.7.4
zipp==3.12.0
